Add : support EVO(except TrueHD); change : some code cleanup and refactoring;
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@4105 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/filters/transform/MPCVideoDec/ffmpeg/ffImgfmt.h
@@ -81,7 +81,7 @@
 
 #include <stddef.h>
 
-typedef int stride_t;
+typedef ptrdiff_t stride_t;
 
 //==================================== xvid4 =====================================
